# Updated perf test cases